change splice generation strategy

+++ db_01_spliceDB.py
@@ -6,37 +6,31 @@ from Bio import SeqIO
 
 
 
-
-def bam2chrbam(raw_path,target_path,chroms):
-    '''this function copies bam file and split it based on chromosome
-    '''
+'''There are two strategies. 1. copy all bam files to a foler and then split them into 
+each scaffold, thus in each scaffold folder, it has all bam files. then run the pipeline'''
+def bam2splitBam(raw_path,target_path,scaffs):
     bams = natsorted(glob.glob(raw_path+'/*.bam'))
-    for bam in bams[3:]:
-        # copy file
+    for bam in bams:
         out = target_path+'/' + os.path.split(bam)[1]
         if not os.path.exists(out):
-            cmd = ('cp {i} {o}').format(i=bam,o=out)
+            cmd = ('cp \'{i}\' {o}').format(i=bam,o=out)
             print cmd
             sarge.run(cmd)
-        if not os.path.exists(out+'.bai'):
-            cmd = ('samtools index {b}').format(b=out)
+            cmd = ('cp \'{i}.bai\' {o}.bai').format(i=bam,o=out)
             print cmd
             sarge.run(cmd)
-        # split bam file
-        bam_folder = target_path+'/'+os.path.split(bam)[1][:-4]
-        if not os.path.exists(bam_folder):
-            os.mkdir(bam_folder)
-        for scaff in chroms:
-            cmd = ('samtools view -bh {i} {c} > {out}').format(i=out,c=scaff,out=bam_folder+'/'+scaff+'.bam')
-            print cmd
-            sarge.run(cmd)
+        for scaff in scaffs:
+            bam_path = target_path + '/bam'+scaff
+            if not os.path.exists(bam_path): os.mkdir(bam_path)
+            out_fn = bam_path+'/'+bam.split('/')[-1]
+            if not os.path.exists(out_fn):
+                cmd = ('samtools view -bh \'{i}\' {c} > {out}').format(i=bam,c=scaff,out=out_fn)
+                print cmd
+                sarge.run(cmd)
         os.remove(out)
-            
+        os.remove(out+'.bai')   
 
-
-
-
-def run_spliceDB(scaff,raw_path,target_path,code_path):
+def run_spliceDB_old(scaff,raw_path,target_path,code_path):
     '''
     * scaff: scaffold 
     * target_path: folder that have all results
@@ -44,9 +38,8 @@ def run_spliceDB(scaff,raw_path,target_path,code_path):
     '''
     # create folder
     bam_path = target_path + '/bam'+scaff
-    if os.path.exists(bam_path): 
-        shutil.rmtree(bam_path)
-    os.mkdir(bam_path)
+    if not os.path.exists(bam_path): 
+        assert False,'file not exitts'
     spl_path = target_path + '/spl'+scaff
     if os.path.exists(spl_path): 
         shutil.rmtree(spl_path)
@@ -57,12 +50,6 @@ def run_spliceDB(scaff,raw_path,target_path,code_path):
     os.mkdir(dna_path)
     fa_res_path = target_path+'/prfa'
     if not os.path.exists(fa_res_path): os.mkdir(fa_res_path)
-    bam2chrbam(raw_path,target_path,[scaff])
-    # copy bam file and extract chromosome specific bam file
-    bams = glob.glob(raw_path+'/*/'+scaff+'.bam')
-    for bam in bams:
-        sarge.run('cp {before} {after}'.format(before=bam,after=bam_path+'/'+bam.split('/')[-2]+'.bam'))
-    # create fa file
     global ref_dic
     fa = dna_path+'/'+scaff+'.fa'
     if not os.path.exists(fa):
@@ -70,7 +57,7 @@ def run_spliceDB(scaff,raw_path,target_path,code_path):
             SeqIO.write(ref_dic[scaff], handle,'fasta')
     # run splice db
     buildsp = code_path + '/buildSpliceGraph.py'
-    sam = "/usr/local/bin" # samtools path
+    sam = "/opt/miniconda2/bin" # samtools path
      
     cmd = 'python {build} -a {bam} -p {spl} -t {sam} -f {fa} -c {c}'.format(build=buildsp,bam=bam_path,spl=spl_path,sam=sam,fa=dna_path,c=code_path)
     sarge.run(cmd)
@@ -80,22 +67,109 @@ def run_spliceDB(scaff,raw_path,target_path,code_path):
         shutil.move(f,fa_res_path)
     # remove those folders
     shutil.rmtree(bam_path)
-    shutil.rmtree(spl_path)
+    # shutil.rmtree(spl_path)
     shutil.rmtree(dna_path)
 
+
+# import time
+# start = time.time()
+
+
+# # raw_path = '/data/shangzhong/GoogleDrive/Research_Project/CHO_Project/Bam_files/Proteogenomics'
+# raw_path = '/media/lewislab/Dropbox (UCSD SBRG)/users/shangzhong/proteogenomics_bam'
+# target_path = '/data/shangzhong/Proteogenomics/Database/Splice_Db'
+# ref_fa = '/data/genome/hamster/picr_old/picr.fa'
+# ref_dic = SeqIO.index(ref_fa,'fasta')
+
+
+# scaffs = ['picr_'+str(n) for n in [0,1]]
+# code_path = '/home/shangzhong/Codes/Proteogenomics/SpliceDB/bin'
+# thread = 9
+
+# # bam2splitBam(raw_path,target_path,list(ref_dic.keys()))
+
+# pool = mp.Pool(processes=thread)
+# for scaff in scaffs:
+#     pool.apply_async(run_spliceDB,args=(scaff,raw_path,target_path,code_path,))
+# pool.close()
+# pool.join()
+
+
+# print time.time() - start
+
+# before this line is code to copy bam to your server first, the following is directly extract chr bams.
+
+
+
+'''The second method is directly extract bam file for each scaffold and then run the pipeline.
+'''
+def bam2splitBam(raw_path,target_path,scaff):
+    bams = natsorted(glob.glob(raw_path+'/*.bam'))
+    for bam in bams:
+        out_fn = target_path+'/'+bam.split('/')[-1]
+        if not os.path.exists(out_fn):
+            cmd = ('samtools view -bh \'{i}\' {c} > {out}').format(i=bam,c=scaff,out=out_fn)
+            print cmd
+            sarge.run(cmd)
+
+
+def run_spliceDB(scaff,raw_path,target_path,code_path):
+    '''
+    * scaff: scaffold 
+    * target_path: folder that have all results
+    * code_path: splice db codes path
+    '''
+    # create folder
+    bam_path = target_path + '/bam_'+scaff
+    if os.path.exists(bam_path): 
+        shutil.rmtree(bam_path)
+    os.mkdir(bam_path)
+    spl_path = target_path + '/spl_'+scaff
+    if os.path.exists(spl_path): 
+        shutil.rmtree(spl_path)
+    os.mkdir(spl_path)
+    dna_path = target_path + '/dna_'+scaff
+    if os.path.exists(dna_path): 
+        shutil.rmtree(dna_path)
+    os.mkdir(dna_path)
+    fa_res_path = target_path+'/prfa'
+    if not os.path.exists(fa_res_path): os.mkdir(fa_res_path)
+    bam2splitBam(raw_path,bam_path,scaff)
+    # create fa file
+    global ref_dic
+    fa = dna_path+'/'+scaff+'.fa'
+    if not os.path.exists(fa):
+        with open(dna_path+'/'+scaff+'.fa','w') as handle:
+            SeqIO.write(ref_dic[scaff], handle,'fasta')
+    # run splice db
+    buildsp = code_path + '/buildSpliceGraph.py'
+    sam = "/opt/miniconda2/bin" # samtools path
+     
+    cmd = 'python {build} -a {bam} -p {spl} -t {sam} -f {fa} -c {c}'.format(build=buildsp,bam=bam_path,spl=spl_path,sam=sam,fa=dna_path,c=code_path)
+    print(cmd)
+    sarge.run(cmd)
+    # move fa file to prfa folder
+    fa = glob.glob(spl_path+'/*.fa')
+    for f in fa:
+        shutil.move(f,fa_res_path)
+    # remove those folders
+    shutil.rmtree(bam_path)
+    # shutil.rmtree(spl_path)
+    shutil.rmtree(dna_path)
 
 import time
 start = time.time()
 
 # bam2chrbam(raw_path,target_path,list(ref_dic.keys()))
 
-# raw_path = '/data/shangzhong/GoogleDrive/Research_Project/CHO_Project/Bam_files/Proteogenomics'
+
 raw_path = '/media/lewislab/Dropbox (UCSD SBRG)/users/shangzhong/proteogenomics_bam'
-target_path = '/data/shangzhong/Proteogenomics/Database/Splice_Db'
+target_path = '/data/shangzhong/Proteogenomics/Database/Splice_Db_9'
+if not os.path.exists(target_path): os.mkdir(target_path)
 ref_fa = '/data/genome/hamster/picr_old/picr.fa'
 ref_dic = SeqIO.index(ref_fa,'fasta')
 
-scaffs = ['picr_'+str(n) for n in range(1829)]
+scaffs = ['picr_'+str(n) for n in [9]]
 code_path = '/home/shangzhong/Codes/Proteogenomics/SpliceDB/bin'
 thread = 9
